refactor: Models 디렉토리 생성 후 이동 CommonWeatherModels.swift 생성 후 각 타입별 세부타입 분리
