--- WeatherForecast/WeatherForecast/CurrentWeather.swift
@@ -1,74 +1,0 @@
-//
-//  CurrentWeather.swift
-//  WeatherForecast
-//
-//  Created by tae hoon park on 2021/09/28.
-//
-
-import Foundation
-
-struct Coord: Decodable {
-    let lon: Double
-    let lat: Double
-}
-
-struct Weather: Decodable {
-    let id: Int
-    let main: String
-    let description: String
-    let icon: String
-}
-
-struct MainWeatherInfo: Decodable {
-    let temp: Double
-    let feelsLike: Double
-    let tempMin: Double
-    let tempMax: Double
-    let pressure: Int
-    let humidity: Int
-    let seaLevel: Int
-    let grndLevel: Int
-    
-    enum CodingKeys: String, CodingKey {
-        case temp, pressure, humidity
-        case feelsLike = "feels_like"
-        case tempMin = "temp_min"
-        case tempMax = "temp_max"
-        case seaLevel = "sea_level"
-        case grndLevel = "grnd_level"
-    }
-}
-
-struct Wind: Decodable {
-    let speed: Double
-    let deg: Int
-    let gust: Double
-}
-
-struct Clouds: Decodable {
-    let all: Int
-}
-
-struct Sys: Decodable {
-    let type: Int
-    let id: Int
-    let country: String
-    let sunrise: Int
-    let sunset: Int
-}
-
-struct CurrentWeather: Decodable {
-    let coord: Coord
-    let weather: [Weather]
-    let base: String
-    let main: MainWeatherInfo
-    let visibility: Int
-    let wind: Wind
-    let clouds: Clouds
-    let dt: Int
-    let sys: Sys
-    let timezone: Int
-    let id: Int
-    let name: String
-    let cod: Int
-}
