+++ WeatherForecast/WeatherForecast/Models/CommonWeatherModels.swift
@@ -1,5 +1,5 @@
 //
-//  FiveDaysWeather.swift
+//  CommonWeatherModel.swift
 //  WeatherForecast
 //
 //  Created by tae hoon park on 2021/09/28.
@@ -7,6 +7,59 @@
 
 import Foundation
 
+// MARK: - For CurrentWeather, FiveDaysWeather
+struct Weather: Decodable {
+    let id: Int
+    let main: String
+    let description: String
+    let icon: String
+}
+
+struct Coord: Decodable {
+    let lon: Double
+    let lat: Double
+}
+
+struct Wind: Decodable {
+    let speed: Double
+    let deg: Int
+    let gust: Double
+}
+
+struct Clouds: Decodable {
+    let all: Int
+}
+
+// MARK: - For CurrentWeather
+struct MainWeatherInfo: Decodable {
+    let temp: Double
+    let feelsLike: Double
+    let tempMin: Double
+    let tempMax: Double
+    let pressure: Int
+    let humidity: Int
+    let seaLevel: Int
+    let grndLevel: Int
+    
+    enum CodingKeys: String, CodingKey {
+        case temp, pressure, humidity
+        case feelsLike = "feels_like"
+        case tempMin = "temp_min"
+        case tempMax = "temp_max"
+        case seaLevel = "sea_level"
+        case grndLevel = "grnd_level"
+    }
+}
+
+struct Sys: Decodable {
+    let type: Int
+    let id: Int
+    let country: String
+    let sunrise: Int
+    let sunset: Int
+}
+
+// MARK: - For FiveDaysWeather
 struct ListMainWeatherInfo: Decodable {
     let temp: Double
     let feelsLike: Double
@@ -49,7 +102,6 @@ struct List: Decodable {
              clouds, wind, visibility, pop, sys
         case dtTxt = "dt_txt"
     }
-    
 }
 
 struct City: Decodable {
@@ -61,12 +113,4 @@ struct City: Decodable {
     let timezone: Int
     let sunrise: Int
     let sunset: Int
-}
-
-struct FiveDaysWeather: Decodable {
-    let cod: Int
-    let message: Int
-    let cnt: Int
-    let list: [List]
-    let city: City
 }
